refactor: move document update effect to own hook
